Rename hedaer file to SliceReader
