+++ src/Archive/VolFile.cpp
@@ -1,5 +1,5 @@
 #include "VolFile.h"
-#include "../Stream/ReaderSlice.h"
+#include "../Stream/SliceReader.h"
 #include "../XFile.h"
 #include <stdexcept>
 #include <algorithm>
