+++ src/Stream/SliceReader.h
